fix: nested tokens under theme
