+++ lib/predicates.ts
@@ -1,12 +1,11 @@
 import type { Token } from 'style-dictionary';
-import { getType } from './sources.ts';
 
 export type Predicate = (token: Token) => boolean;
 
 const and = (p: Predicate, q: Predicate): Predicate => x => p(x) && q(x);
 
 export const isColor: Predicate = token =>
-  getType(token) === 'color';
+  token.$type === 'color';
 
 export const isBreakpoint: Predicate = token =>
   token.path.includes('breakpoint');
@@ -18,19 +17,19 @@ export const isTablet: Predicate = token =>
   token.path.includes('tablet');
 
 export const isShadow: Predicate = token =>
-  getType(token) === 'shadow';
+  token.$type === 'shadow';
 
 export const isCubicBezier: Predicate = token =>
-  getType(token) === 'cubicBezier';
+  token.$type === 'cubicBezier';
 
 export const isFontFamily: Predicate = token =>
-  getType(token) === 'fontFamily';
+  token.$type === 'fontFamily';
 
 export const isFontWeight: Predicate = token =>
-  getType(token) === 'fontWeight';
+  token.$type === 'fontWeight';
 
 export const isMediaQuery: Predicate = token =>
-  getType(token) === 'mediaQuery';
+  token.$type === 'mediaQuery';
 
 export const hasExtensions: Predicate = token =>
   !!token.$extensions;
