+++ package.json
@@ -103,7 +103,7 @@
       ]
     },
     "11ty": {
-      "command": "eleventy",
+      "command": "NODE_OPTIONS=\"--import tsx\" eleventy --config=eleventy.config.ts",
       "dependencies": [
         "style-dictionary",
         "compile"
@@ -160,7 +160,7 @@
     }
   },
   "devDependencies": {
-    "@11ty/eleventy": "^2.0.1",
+    "@11ty/eleventy": "^3.0.0-alpha.18 || ^3.0.0",
     "@changesets/cli": "^2.27.7",
     "@eslint/eslintrc": "^3.1.0",
     "@eslint/js": "^9.9.1",
@@ -177,6 +177,7 @@
     "tap-difflet": "^0.7.2",
     "tape-describe": "^1.0.3",
     "tape-es": "^1.2.17",
+    "tsx": "^4.19.0",
     "typescript": "^5.5.4",
     "vsce": "^2.15.0",
     "wireit": "^0.14.8",
