+++ tokens/color/brand.md
@@ -1,0 +1,6 @@
+The primary brand color is red, it brings boldness and helps guide the eye or highlight important information.
+
+To learn more about how the red color is used, visit the [Brand standards][standards] page.
+
+[standards]: https://www.redhat.com/en/about/brand/standards/color
+
